backend/requirements.txt
fastapi==0.115.0
uvicorn==0.30.0
sqlalchemy==2.0.35
pydantic[email]==2.9.0
passlib[bcrypt]==1.7.4
bcrypt==4.0.1
python-jose[cryptography]==3.3.0
python-dotenv==1.0.1
httpx==0.27.0
apscheduler==3.10.4
slowapi==0.1.9
